Abstract out BundleElectionSolver from BundleProducer
A pure refactoring without any logical changes.
